# `<bit>`: Improve `has_single_bit()` codegen