--- a/benchmarks/src/has_single_bit.cpp
+++ b/benchmarks/src/has_single_bit.cpp
@@ -8,11 +8,12 @@
 #include <utility.hpp>
 
 template <typename T>
-static void has_single_bit_if(benchmark::State& state) {
+void has_single_bit_if(benchmark::State& state) {
     const auto random_v = random_vector<T>(8);
     for (auto _ : state) {
         benchmark::DoNotOptimize(random_v);
-        int count_true = 0, count_false = 0;
+        unsigned count_true  = 0;
+        unsigned count_false = 0;
         for (const auto& x : random_v) {
             if (std::has_single_bit(x)) {
                 benchmark::DoNotOptimize(count_true++);
@@ -24,12 +25,12 @@ static void has_single_bit_if(benchmark::State& state) {
 }
 
 template <typename T>
-static void has_single_bit(benchmark::State& state) {
+void has_single_bit(benchmark::State& state) {
     const auto random_v = random_vector<T>(8);
     assert(random_v.size() % 4 == 0);
     for (auto _ : state) {
         benchmark::DoNotOptimize(random_v);
-        int r = 0;
+        unsigned r = 0;
         for (size_t i = 0; i < random_v.size(); i += 4) {
             r += std::has_single_bit(random_v[i + 0]);
             r += std::has_single_bit(random_v[i + 1]);