code avoids branches, provides 20% speed up on x64 targets.

**x64 results**
```
---------------------------------------
Benchmark                          Time
---------------------------------------
Rand16_has_single_bit            284 ns
Rand16_has_single_bit_new        232 ns
Rand32_has_single_bit            306 ns
Rand32_has_single_bit_new        246 ns
Rand64_has_single_bit            289 ns
Rand64_has_single_bit_new        233 ns
```


partially addresses #5359