@pps83 2025-03-27
seems like code is better for x86/x64/arm/arm64 : https://godbolt.org/z/fjshqMf4d

the only potential drawback: old code would check for _Val = 0 and jump. New code has no jumps (except on arm)