@AlexGuteniev 2025-03-27
> (except on arm)

Note that only 64-bit arm is an important optimization target.
It is referred to as ARM64 here and AArch64 somewhere else.

32-bit arm is abandoned as optimization target.