@pps83 2025-03-27
> 32-bit arm is abandoned as optimization target.

Imo for mobiles it should still be relevant. More relevant that x86 for x64 CPUs. In any case, 32bit arm was the same