@AlexGuteniev 2025-03-27
i5-1235U

x64

P-core

Benchmark                    |  Before  |    After  
-----------------------------|----------|-----------
has_single_bit_if<uint16_t>  | 13,5 ns  |  12,4 ns  
has_single_bit_if<uint32_t>  | 14,4 ns  |  11,1 ns  
has_single_bit_if<uint64_t>  | 13,8 ns  |  13,1 ns  
has_single_bit<uint16_t>     | 5,96 ns  |  4,53 ns  
has_single_bit<uint32_t>     | 5,53 ns  |  5,03 ns  
has_single_bit<uint64_t>     | 5,22 ns  |  5,02 ns  

E-core

Benchmark                    |  Before  |   After  
-----------------------------|----------|-----------
has_single_bit_if<uint16_t>  | 19,7 ns  |  17.7 ns  
has_single_bit_if<uint32_t>  | 19,7 ns  |  17.7 ns  
has_single_bit_if<uint64_t>  | 19,7 ns  |  19.9 ns  
has_single_bit<uint16_t>     | 12,5 ns  |  9.35 ns  
has_single_bit<uint32_t>     | 12,0 ns  |  9.33 ns  
has_single_bit<uint64_t>     | 13,2 ns  |  9.37 ns