@pps83 2025-03-27
Can you please try this local change also:

```c++
_NODISCARD constexpr bool has_single_bit(const _Ty _Val) noexcept {
#if 1 // _POPCNT_INTRINSICS_ALWAYS_AVAILABLE
#if _HAS_CXX20
    if (_STD is_constant_evaluated())
        return 1 == _Popcount_fallback(_Val);
#endif // _HAS_CXX20
    return 1 == _Unchecked_popcount(_Val);
#else
    return (_Val ^ (_Val - 1)) > _Val - 1;
#endif
}
```