@AlexGuteniev 2025-03-27
Turns out this particular benchmark is very much affected by #4496 and `popcnt` version is slower. 

That's why I've hidden the previous results, and obtained the updated results by changing also this line to say `Release` instead of `RelWithDebInfo`

https://github.com/microsoft/STL/blob/main/benchmarks/CMakeLists.txt#L62

i5-1235U

x64

P-core

Benchmark                    |  Before  |    After  | popcnt
-----------------------------|----------|-----------|---------
has_single_bit_if<uint16_t>  | 11.5 ns  |  10.7 ns  | 6.68 ns 
has_single_bit_if<uint32_t>  | 11.3 ns  |  11.3 ns  | 10.6 ns 
has_single_bit_if<uint64_t>  | 6.98 ns  |  11.7 ns  | 6.57 ns 
has_single_bit<uint16_t>     | 6.36 ns  |  4.51 ns  | 3.51 ns 
has_single_bit<uint32_t>     | 6.08 ns  |  4.71 ns  | 3.52 ns 
has_single_bit<uint64_t>     | 5.27 ns  |  4.71 ns  | 3.53 ns 

E-core

Benchmark                    |  Before  |   After   | popcnt  
-----------------------------|----------|-----------|---------
has_single_bit_if<uint16_t>  | 16.2 ns  |  14.1 ns  | 14.1 ns 
has_single_bit_if<uint32_t>  | 16.7 ns  |  14.0 ns  | 14.0 ns 
has_single_bit_if<uint64_t>  | 16.2 ns  |  16.2 ns  | 14.0 ns 
has_single_bit<uint16_t>     | 12.2 ns  |  9.07 ns  | 7.24 ns 
has_single_bit<uint32_t>     | 11.9 ns  |  9.07 ns  | 7.24 ns 
has_single_bit<uint64_t>     | 11.6 ns  |  9.07 ns  | 7.25 ns 

Apparently `popcnt` is a good idea. Even for the "if" case.
----

#4496 wasn't a severe obstacle for most previous optimization, but at such micro level it is really disrupting, we'll need to get into that.