@AlexGuteniev 2025-03-27
> Can you please try this local change also:

Before you go with this in the PR:
 * We don't compare like `1 == function()` in new code, although there are some relic precedents. In new code, we use natural order `function() == 1`
 * Always control flow braces `{}` even around single statement in `if`
 * You don't need to check `#if _HAS_CXX20`. It is C++20 header, and the check is already there at the very beginning for the whole content