@pps83 2025-03-27
> Before you go with this in the PR...

This PR doesn't have popcnt change. The popcnt change doesn't use dispatcher, it's the unchecked version that does the `popcnt` opcode directly and it should only be used when building for AVX+. The default (version with this PR) is better than using dispatcher imo.

I can add popcnt it as a separate PR, i think it's better  that way.