@StephanTLavavej 2025-03-27
For the record, ARM32 performance is indeed utterly irrelevant. Windows ripped out their ARM32 build, and will even be dropping support for ARM32 binaries running on ARM64 machines, see https://learn.microsoft.com/en-us/windows/arm/arm32-to-arm64 .

ARM32 still needs to behave correctly (until DevDiv/MSVC drops support for targeting it, which I am constantly asking about), but we don't want to waste any more time on it.