@StephanTLavavej 2025-04-01
Thanks! I pushed some changes to the benchmark for consistency with repo conventions.